+++ tests/testthat/test-gridifyLayout.R
@@ -8,6 +8,7 @@ test_that("gridifyLayout can be created with a proper input", {
       widths = grid::unit(1, "npc"),
       margin = grid::unit(c(1, 1, 1, 1), "npc"),
       global_gpar = grid::gpar(),
+      background = grid::get.gpar()$fill,
       adjust_height = FALSE,
       object = gridifyObject(1, 1),
       cells = gridifyCells(sth = gridifyCell(1, 1))
@@ -23,18 +24,31 @@ test_that(
   {
     class_spec <- getSlots("gridifyLayout")
 
-    expect_identical(class_spec, c(
-      nrow = "numeric", ncol = "numeric", heights = "unitOrSimpleUnit",
-      widths = "unitOrSimpleUnit", margin = "unitOrSimpleUnit", global_gpar = "gpar",
-      adjust_height = "logical", object = "gridifyObject", cells = "gridifyCells"
-    ))
+    expect_identical(
+      class_spec,
+      c(
+        nrow = "numeric",
+        ncol = "numeric",
+        heights = "unitOrSimpleUnit",
+        widths = "unitOrSimpleUnit",
+        margin = "unitOrSimpleUnit",
+        global_gpar = "gpar",
+        background = "character",
+        adjust_height = "logical",
+        object = "gridifyObject",
+        cells = "gridifyCells"
+      )
+    )
   }
 )
 
 
 test_that("unitOrSimpleUnit is unit or simpleUnit", {
   expect_true(methods::is(grid::unit(1, "npc"), "unitOrSimpleUnit"))
-  expect_true(methods::is(grid::unit(c(1, 2), c("npc", "cm")), "unitOrSimpleUnit"))
+  expect_true(methods::is(
+    grid::unit(c(1, 2), c("npc", "cm")),
+    "unitOrSimpleUnit"
+  ))
 })
 
 test_that("gridifyLayout setValidity tests", {
@@ -47,6 +61,7 @@ test_that("gridifyLayout setValidity tests", {
       widths = grid::unit(1, "npc"),
       margin = grid::unit(c(1, 1, 1), "cm"),
       global_gpar = grid::gpar(),
+      background = grid::get.gpar()$fill,
       adjust_height = FALSE,
       object = gridifyObject(1, 1),
       cells = gridifyCells(sth = gridifyCell(1, 1))
@@ -63,6 +78,7 @@ test_that("gridifyLayout setValidity tests", {
       widths = grid::unit(1, "npc"),
       margin = grid::unit(c(1, 1, 1, 1), "cm"),
       global_gpar = grid::gpar(),
+      background = grid::get.gpar()$fill,
       adjust_height = FALSE,
       object = gridifyObject(1, 1),
       cells = gridifyCells(sth = gridifyCell(1, 1))
@@ -79,6 +95,7 @@ test_that("gridifyLayout setValidity tests", {
       widths = grid::unit(1, "npc"),
       margin = grid::unit(c(1, 1, 1, 1), "cm"),
       global_gpar = grid::gpar(),
+      background = grid::get.gpar()$fill,
       adjust_height = FALSE,
       object = gridifyObject(1, 1),
       cells = gridifyCells(sth = gridifyCell(1, 1))
@@ -95,9 +112,13 @@ test_that("gridifyLayout setValidity tests", {
       widths = grid::unit(c(1, 1), "npc"),
       margin = grid::unit(c(1, 1, 1, 1), "cm"),
       global_gpar = grid::gpar(),
+      background = grid::get.gpar()$fill,
       adjust_height = FALSE,
       object = gridifyObject(1, 1),
-      cells = gridifyCells(a = gridifyCell(row = 1, col = 3), b = gridifyCell(row = 2, col = 2))
+      cells = gridifyCells(
+        a = gridifyCell(row = 1, col = 3),
+        b = gridifyCell(row = 2, col = 2)
+      )
     ),
     "All cells cols have to be less or equal to ncol"
   )
@@ -111,11 +132,15 @@ test_that("gridifyLayout setValidity tests", {
       widths = grid::unit(c(1, 1), "npc"),
       margin = grid::unit(c(1, 1, 1, 1), "cm"),
       global_gpar = grid::gpar(),
+      background = grid::get.gpar()$fill,
       adjust_height = FALSE,
       object = gridifyObject(10, 1),
-      cells = gridifyCells(a = gridifyCell(row = 1, col = 2), b = gridifyCell(row = 2, col = 2))
+      cells = gridifyCells(
+        a = gridifyCell(row = 1, col = 2),
+        b = gridifyCell(row = 2, col = 2)
+      )
     ),
-    "less or equal to nrow: 2"
+    "row value has to be less or equal to nrow: 2"
   )
 
   expect_error(
@@ -127,11 +152,15 @@ test_that("gridifyLayout setValidity tests", {
       widths = grid::unit(c(1, 1), "npc"),
       margin = grid::unit(c(1, 1, 1, 1), "cm"),
       global_gpar = grid::gpar(),
+      background = grid::get.gpar()$fill,
       adjust_height = FALSE,
       object = gridifyObject(1, 10),
-      cells = gridifyCells(a = gridifyCell(row = 1, col = 2), b = gridifyCell(row = 2, col = 2))
+      cells = gridifyCells(
+        a = gridifyCell(row = 1, col = 2),
+        b = gridifyCell(row = 2, col = 2)
+      )
     ),
-    "less or equal to ncol: 2"
+    "col value has to be less or equal to ncol: 2"
   )
 
   expect_warning(
@@ -143,9 +172,13 @@ test_that("gridifyLayout setValidity tests", {
       widths = grid::unit(c(1, 1, 1), "npc"),
       margin = grid::unit(c(1, 1, 1, 1), "cm"),
       global_gpar = grid::gpar(),
+      background = grid::get.gpar()$fill,
       adjust_height = FALSE,
       object = gridifyObject(2, 2),
-      cells = gridifyCells(a = gridifyCell(row = 1:2, col = 2), b = gridifyCell(row = 2, col = 2))
+      cells = gridifyCells(
+        a = gridifyCell(row = 1:2, col = 2),
+        b = gridifyCell(row = 2, col = 2)
+      )
     ),
     "Overlapping cells detected at positions: 2-2"
   )
@@ -159,9 +192,13 @@ test_that("gridifyLayout setValidity tests", {
       widths = grid::unit(c(1, 1, 1), "npc"),
       margin = grid::unit(c(1, 1, 1, 1), "cm"),
       global_gpar = grid::gpar(),
+      background = grid::get.gpar()$fill,
       adjust_height = FALSE,
       object = gridifyObject(2, 2),
-      cells = gridifyCells(a = gridifyCell(row = 1:3, col = 3), b = gridifyCell(row = 3, col = 3))
+      cells = gridifyCells(
+        a = gridifyCell(row = 1:3, col = 3),
+        b = gridifyCell(row = 3, col = 3)
+      )
     ),
     "Overlapping cells detected at positions: 3-3"
   )
@@ -175,9 +212,13 @@ test_that("gridifyLayout setValidity tests", {
       widths = grid::unit(c(1, 1, 1), "npc"),
       margin = grid::unit(c(1, 1, 1, 1), "cm"),
       global_gpar = grid::gpar(),
+      background = grid::get.gpar()$fill,
       adjust_height = FALSE,
       object = gridifyObject(2, 2),
-      cells = gridifyCells(a = gridifyCell(row = 1:2, col = 3), b = gridifyCell(row = 3, col = 3))
+      cells = gridifyCells(
+        a = gridifyCell(row = 1:2, col = 3),
+        b = gridifyCell(row = 3, col = 3)
+      )
     ),
     NA
   )