+++ NEWS.md
@@ -1,3 +1,8 @@
+# gridify 0.7.3
+
+* Improved `show_spec` method for `gridifyLayout`.
+* New feature of background colour for layouts, by the new `background` argument.
+
 # gridify 0.7.2
 
 * First open-source release.